requirements.txt
Django==4.2.7
google-generativeai==0.3.1
python-decouple==3.8
Pillow==10.1.0